Turn `SourceName` into an interface
